feat: add carrier console UI, WebRTC secure-context guard, and ORT env scripts

+++ crates/pjsip/src/endpoint.rs
@@ -197,7 +197,7 @@ impl PjEndpoint {
     fn start_udp_transport(&mut self) -> Result<()> {
         let mut addr: pjsip_sys::pj_sockaddr_in = unsafe { std::mem::zeroed() };
         // sin_family is pj_uint8_t (u8) — PJ_AF_INET = 2 (AF_INET)
-        addr.sin_family = libc::AF_INET as u8;
+        addr.sin_family = libc::AF_INET as u16;
         addr.sin_port = self.config.port.to_be();
         // 0.0.0.0 = INADDR_ANY
         addr.sin_addr.s_addr = 0;
@@ -218,7 +218,7 @@ impl PjEndpoint {
 
     fn start_tcp_transport(&mut self) -> Result<()> {
         let mut addr: pjsip_sys::pj_sockaddr_in = unsafe { std::mem::zeroed() };
-        addr.sin_family = libc::AF_INET as u8;
+        addr.sin_family = libc::AF_INET as u16;
         addr.sin_port = self.config.port.to_be();
         addr.sin_addr.s_addr = 0;
 
@@ -259,7 +259,7 @@ impl PjEndpoint {
             unsafe { pjsip_sys::pj_str(key_cstr.as_ptr() as *mut _) };
 
         let mut addr: pjsip_sys::pj_sockaddr_in = unsafe { std::mem::zeroed() };
-        addr.sin_family = libc::AF_INET as u8;
+        addr.sin_family = libc::AF_INET as u16;
         addr.sin_port = self.config.port.to_be();
         addr.sin_addr.s_addr = 0;
